common: refactor large source
